requirements.txt
Flask
gunicorn
speedtest-cli
fpdf